refactor(pages): update navbar and footer imports
