+++ src/pages/TaiexPage/Taiex.js
@@ -1,8 +1,8 @@
 import { useCollection } from "../../hooks/useCollection"
 
 // components
-import Navbar from "../../components/navbar/Navbar"
-import Footer from '../../components/footer/Footer'
+import Navbar from "../../components/layout/Navbar"
+import Footer from '../../components/layout/Footer'
 import { TaiexSVG } from "./TaiexSVG"
 import loadingGif from '../../assets/loading.gif'
